perf: hardening concurrencia — pool, timeouts, cache JWT y paginacion
- HikariCP: pool 4→10, minimum-idle 1→2 (AWS RDS soporta 60+)
- StockSseRegistry: @SchedulerLock en heartbeat() — un solo pod en multi-pod
- AssistantMetricsService: @SchedulerLock en logStats() cada 6h
- ExtraccionService: scraping secuencial→paralelo con CompletableFuture + timeout 12s
- WhatsAppService: RestTemplate con connect 5s / read 10s (antes sin timeout)
- JwtRequestFilter: cache Caffeine TTL 30s para UserDetails (evita query a BD por request)
- VentaController /clientes: LIMIT 500 via findTop500 (antes cargaba toda la tabla)
- BlogController /admin: LIMIT 100 via findTop100ByOrderByFechaCreacionDesc
- ConvenioController /admin: LIMIT 100 via findTop100ByOrderByFechaRegistroDesc

+++ Hot_click_outlet/src/main/java/com/hotclick/service/PaymentService.java
@@ -177,7 +177,7 @@ public class PaymentService {
         pedido.setAplicaImpuesto(false);
         if (subtotal > 0) {
             pedido.setMargenGananciaPedido(
-                BigDecimal.valueOf(subtotal - costoTotal)
+                BigDecimal.valueOf((long) subtotal - costoTotal)
                     .divide(BigDecimal.valueOf(subtotal), 4, RoundingMode.HALF_UP)
                     .multiply(BigDecimal.valueOf(100)));
         }
