+++ Hot_click_outlet/src/main/java/com/hotclick/service/SinpeService.java
@@ -120,7 +120,7 @@ public class SinpeService {
         pedido.setAplicaImpuesto(false);
         if (subtotal > 0) {
             pedido.setMargenGananciaPedido(
-                BigDecimal.valueOf(subtotal - costoTotal)
+                BigDecimal.valueOf((long) subtotal - costoTotal)
                     .divide(BigDecimal.valueOf(subtotal), 4, RoundingMode.HALF_UP)
                     .multiply(BigDecimal.valueOf(100)));
         }
